R6 class that provides summary statistics for batch calculations.

The `PivotBatchStatistics` class contains a set of summary statistics that track how many calculations are batch compatible/incompatible.

Format

R6Class object.

Active bindings

asString

A text description of the batch statistics.

Methods

Public methods

PivotBatchStatistics$new()

PivotBatchStatistics$reset()

PivotBatchStatistics$incrementNoData()

PivotBatchStatistics$incrementCompatible()

PivotBatchStatistics$incrementIncompatible()

PivotBatchStatistics$clone()

Method new()

Create a new `PivotBatchStatistics` object.

Usage

PivotBatchStatistics$new(parentPivot = NULL)

Arguments

parentPivot

The pivot table that this `PivotBatchStatistics` instance belongs to.

Returns

A new `PivotBatchStatistics` object.

Method reset()

Clear the batch statistics.

Usage

PivotBatchStatistics$reset()

Returns

No return value.

Method incrementNoData()

Increment the noData count for a batch.

Usage

PivotBatchStatistics$incrementNoData(
  calculationName = NULL,
  calculationGroupName = NULL
)

Arguments

calculationName

The name of the calculation to increment the count for.

calculationGroupName

The name of the calculation group for the calculation.

Returns

No return value.

Method incrementCompatible()

Increment the compatible count for a batch.

Usage

PivotBatchStatistics$incrementCompatible(
  calculationName = NULL,
  calculationGroupName = NULL
)

Arguments

calculationName

The name of the calculation to increment the count for.

calculationGroupName

The name of the calculation group for the calculation.

Returns

No return value.

Method incrementIncompatible()

Increment the incompatible count for a batch.

Usage

PivotBatchStatistics$incrementIncompatible(
  calculationName = NULL,
  calculationGroupName = NULL
)

Arguments

calculationName

The name of the calculation to increment the count for.

calculationGroupName

The name of the calculation group for the calculation.

Returns

No return value.

Method clone()

The objects of this class are cloneable with this method.

Usage

PivotBatchStatistics$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# This class should only be created by the pivot table.
# It is not intended to be created outside of the pivot table.