A class that represents a cell in a pivot table

The PivotCell class represents a cell in the body of a pivot table (i.e. not a row/column heading, rather a cell typically containing a numerical value).

Format

R6Class object.

Value

Object of R6Class with properties and methods that define a single pivot table cell

Fields

parentPivot

Owning pivot table.

rowNumber

The row number of the cell. 1 = the first (i.e. top) data row.

columnNumber

The column number of the cell. 1 = the first (i.e. leftmost) data column.

rowLeafGroup

The row data group linked to this row.

columnLeafGroup

The column data group linked to this column.

calculationName

The name of the calculation that is displayed in the cell.

calculationGroupName

The name of the calculation group that owns the above calculation.

isEmpty

Indicates whether this cell contains no data (e.g. if it is part of a header / outline row).

rowFilters

The data filters applied to this cell from the row headings.

columnFilters

The data filters applied to this cell from the column headings.

rowColFilters

The data filters applied to this cell from both the row and column headings.

calculationFilters

The data filters applied to this cell from the calculation definition.

workingData

The data filters and batchNames used applied when running calculations (including the filters needed for base calculations when calculation type="calculation").

evaluationFilters

The final and actual data filters used in the calculation of the cell value (e.g. custom calculation functions can override the working filters).

isTotal

Whether this cell is a total cell.

rawValue

The numerical calculation result.

formattedValue

The formatted calculation result (i.e. character data type).

baseStyleName

The name of the style applied to this cell (a character value). The style must exist in the PivotStyles object associated with the PivotTable.

style

A PivotStyle object that can apply overrides to the base style for this cell.

Methods

Documentation

For more complete explanations and examples please see the extensive vignettes supplied with this package.

new(...)

Create a new pivot table cell, specifying the field values documented above.

setStyling(styleDeclarations=NULL))

Used to set style declarations.

getCopy())

Get a copy of this cell.

asList())

Get a list representation of this cell

asJSON()

Get a JSON representation of this cell

Examples

# This class should only be created by the pivot table.
# It is not intended to be created outside of the pivot table.