R6 class that renders a pivot table in HTML.

The `PivotHtmlRenderer` class creates a HTML representation of a pivot table.

Format

R6Class object.

Methods

Public methods

PivotHtmlRenderer$new()

PivotHtmlRenderer$clearIsRenderedFlags()

PivotHtmlRenderer$getTableHtml()

PivotHtmlRenderer$clone()

Method new()

Create a new `PivotHtmlRenderer` object.

Usage

PivotHtmlRenderer$new(parentPivot)

Arguments

parentPivot

The pivot table that this `PivotHtmlRenderer` instance belongs to.

Returns

A new `PivotHtmlRenderer` object.

Method clearIsRenderedFlags()

An internal method used when rendering a pivot table to HTML. Clear the IsRendered flags that exist on the `PivotDataGroup` class.

Usage

PivotHtmlRenderer$clearIsRenderedFlags()

Returns

No return value.

Method getTableHtml()

Generate a HTML representation of the pivot table, optionally including additional detail for debugging purposes.

Usage

PivotHtmlRenderer$getTableHtml(
  styleNamePrefix = NULL,
  includeHeaderValues = FALSE,
  includeRCFilters = FALSE,
  includeCalculationFilters = FALSE,
  includeWorkingData = FALSE,
  includeEvaluationFilters = FALSE,
  includeCalculationNames = FALSE,
  includeRawValue = FALSE,
  includeTotalInfo = FALSE,
  exportOptions = NULL,
  showRowGroupHeaders = FALSE
)

Arguments

styleNamePrefix

A character variable specifying a prefix for all named CSS styles, to avoid style name collisions where multiple pivot tables exist.

includeHeaderValues

Default `FALSE`, specify `TRUE` to render this debug information.

includeRCFilters

Default `FALSE`, specify `TRUE` to render this debug information.

includeCalculationFilters

Default `FALSE`, specify `TRUE` to render this debug information.

includeWorkingData

Default `FALSE`, specify `TRUE` to render this debug information.

includeEvaluationFilters

Default `FALSE`, specify `TRUE` to render this debug information.

includeCalculationNames

Default `FALSE`, specify `TRUE` to render this debug information.

includeRawValue

Default `FALSE`, specify `TRUE` to render this debug information.

includeTotalInfo

Default `FALSE`, specify `TRUE` to render this debug information.

exportOptions

A list of additional export options - see the "A1. Appendix" for details.

showRowGroupHeaders

Default `FALSE`, specify `TRUE` to render the row group headings. See the "Data Groups" vignette for details.

Returns

A list containing HTML tags from the `htmltools` package. Convert this to a character variable using `as.character()`.

Method clone()

The objects of this class are cloneable with this method.

Usage

PivotHtmlRenderer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# This class should only be created by the pivot table.
# It is not intended to be created outside of the pivot table.